+++ src/commands.js
@@ -117,20 +117,19 @@ commands:
           ...    --reveal                show sensitive information
 
    === service manager (svm) ===
-~  svml   --svm-list [TENANT_ID]                                  list all managed service instances and binding
-~  svmll  --svm-long-list [TENANT_ID]                             long list all managed service instances and bindings
-          --svm-repair-bindings SERVICE_PLAN [PARAMS]             repair missing and ambivalent service bindings
-          --svm-fresh-bindings SERVICE_PLAN TENANT_ID [PARAMS]    create new service bindings
-          --svm-refresh-bindings SERVICE_PLAN TENANT_ID [PARAMS]  delete and recreate service bindings
-*         --svm-delete-bindings SERVICE_PLAN TENANT_ID            delete service bindings
-*         --svm-delete SERVICE_PLAN TENANT_ID                     delete service instances and bindings
-          ...    SERVICE_PLAN                                     filter for service plan with "offering:plan"
-                                                                    or "all-services" for all
-          ...    TENANT_ID                                        filter for tenant id or "all-tenants" for all
-          ...    [PARAMS]                                         create binding with custom parameters
-          ...    --json                                           list in json
-          ...    --time                                           list includes timestamps
-          ...    --reveal                                         show sensitive information
+~  svml   --svm-list [TENANT_ID]                                      list all service instances and bindings
+~  svmll  --svm-long-list [TENANT_ID]                                 long list all service instances and bindings
+          --svm-make-bindings-single SERVICE_PLAN TENANT_ID [PARAMS]  make service bindings 1-to-1
+          --svm-make-bindings-double SERVICE_PLAN TENANT_ID [PARAMS]  make service bindings 1-to-2
+*         --svm-delete-bindings SERVICE_PLAN TENANT_ID                delete service bindings
+*         --svm-delete SERVICE_PLAN TENANT_ID                         delete service instances and bindings
+          ...    SERVICE_PLAN                                         filter for service plan with "offering:plan"
+                                                                        or "all-services" for all
+          ...    TENANT_ID                                            filter for tenant id or "all-tenants" for all
+          ...    [PARAMS]                                             create binding with custom parameters
+          ...    --json                                               list in json
+          ...    --time                                               list includes timestamps
+          ...    --reveal                                             show sensitive information
 
    === server diagnostic (srv) ===
 ~  srvenv  --server-env [APP_NAME]                            dump system environment
@@ -357,25 +356,18 @@ const APP_COMMAND_INFOS = Object.freeze({
     callback: svm.serviceManagerLongList,
     readonly: true,
   },
-  SVM_REPAIR_BINDINGS: {
-    commandVariants: ["--svm-repair-bindings"],
-    requiredPassArgs: [PASS_ARG.SERVICE_PLAN],
-    optionalPassArgs: [PASS_ARG.PARAMS],
-    callback: svm.serviceManagerRepairBindings,
-    useCache: false,
-  },
-  SVM_FRESH_BINDINGS: {
-    commandVariants: ["--svm-fresh-bindings"],
+  SVM_MAKE_BINDINGS_SINGLE: {
+    commandVariants: ["--svm-make-bindings-single"],
     requiredPassArgs: [PASS_ARG.SERVICE_PLAN, PASS_ARG.TENANT_ID],
     optionalPassArgs: [PASS_ARG.PARAMS],
-    callback: svm.serviceManagerFreshBindings,
+    callback: svm.serviceManagerMakeBindingsSingle,
     useCache: false,
   },
-  SVM_REFRESH_BINDINGS: {
-    commandVariants: ["--svm-refresh-bindings"],
+  SVM_MAKE_BINDINGS_DOUBLE: {
+    commandVariants: ["--svm-make-bindings-double"],
     requiredPassArgs: [PASS_ARG.SERVICE_PLAN, PASS_ARG.TENANT_ID],
     optionalPassArgs: [PASS_ARG.PARAMS],
-    callback: svm.serviceManagerRefreshBindings,
+    callback: svm.serviceManagerMakeBindingsDouble,
     useCache: false,
   },
   SVM_DELETE_BINDINGS: {
@@ -391,6 +383,30 @@ const APP_COMMAND_INFOS = Object.freeze({
     callback: svm.serviceManagerDeleteInstancesAndBindings,
     useCache: false,
     danger: true,
+  },
+  SVM_REPAIR_BINDINGS: {
+    deprecated: true,
+    commandVariants: ["--svm-repair-bindings"],
+    requiredPassArgs: [PASS_ARG.SERVICE_PLAN],
+    optionalPassArgs: [PASS_ARG.PARAMS],
+    callback: svm.serviceManagerRepairBindingsDeprecated,
+    useCache: false,
+  },
+  SVM_FRESH_BINDINGS: {
+    deprecated: true,
+    commandVariants: ["--svm-fresh-bindings"],
+    requiredPassArgs: [PASS_ARG.SERVICE_PLAN, PASS_ARG.TENANT_ID],
+    optionalPassArgs: [PASS_ARG.PARAMS],
+    callback: svm.serviceManagerFreshBindingsDeprecated,
+    useCache: false,
+  },
+  SVM_REFRESH_BINDINGS: {
+    deprecated: true,
+    commandVariants: ["--svm-refresh-bindings"],
+    requiredPassArgs: [PASS_ARG.SERVICE_PLAN, PASS_ARG.TENANT_ID],
+    optionalPassArgs: [PASS_ARG.PARAMS],
+    callback: svm.serviceManagerRefreshBindingsDeprecated,
+    useCache: false,
   },
 
   SRV_ENVIRONMENT: {